requirements.txt
django-admin-rangefilter
# RESTFRAMEWORK
djangorestframework
markdown
django-filter
